+++ examples/14_sizing_modes/run_size_modes.py
@@ -2,7 +2,7 @@ from h2integrate.core.h2integrate_model import H2IntegrateModel
 
 
 # Create a H2Integrate model
-model = H2IntegrateModel("14_size_mode.yaml")
+model = H2IntegrateModel("22_size_mode_iterative.yaml")
 model.technology_config["technologies"]["electrolyzer"]["model_inputs"]["performance_parameters"][
     "sizing"
 ] = {
@@ -10,10 +10,7 @@ model.technology_config["technologies"]["electrolyzer"]["model_inputs"]["perform
     "resize_by_flow": "hydrogen",
     "max_commodity_ratio": 1.0,
 }
-model.plant_config["technology_interconnections"].append(
-    ["ammonia", "electrolyzer", "max_hydrogen_capacity"]
-)
-model.setup()
+
 # Run the model
 model.run()